--- src/tutorials.rs
@@ -1,20 +1,0 @@
-//! The quicksilver tutorials, generated through Rustdoc
-//!
-//! While this isn't a traditional way of hosting a tutorial, Rustdoc ensures that all the code in
-//! the tutorials is checked when CI runs, keeping it nice and up-to-date.
-//!
-//! Before you jump into the tutorials below, make sure your development environment is ready. If
-//! you're just targeting desktop, all you need is the latest stable Rust. If you're targeting the
-//! web, make sure you install cargo-web first (cargo install -f cargo-web).
-
-pub mod _01_basic;
-pub mod _02_drawing;
-pub mod _03_input;
-pub mod _04_lifecycle;
-pub mod _05_images;
-pub mod _06_asset_combinators;
-pub mod _07_font;
-pub mod _08_sound;
-pub mod _09_animations_and_atlases;
-pub mod _10_views;
-pub mod _11_mesh;
